// src/@type/failurereport.d.ts
export type failuredatas={
    id:string,
    api:string,
                call_method: string,
                params: string,
                datetime:string,
                api_response:string,
                response_time:string,
                response_status:string,
                status:string
}